requirements.txt
fastapi
uvicorn
pillow
rembg
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator